edit: 마이페이지 코드 전체적으로 수정
