+++ src/common/api.ts
@@ -1,5 +1,13 @@
 import axios from 'axios';
-import { doc, getDoc, setDoc, updateDoc } from 'firebase/firestore';
+import {
+  doc,
+  getDoc,
+  setDoc,
+  updateDoc,
+  collection,
+  query,
+  getDocs,
+} from 'firebase/firestore';
 import { db } from './firebase';
 
 const BASE_URL = 'https://api.odcloud.kr/api/ApplyhomeInfoDetailSvc/v1';
@@ -45,4 +53,20 @@ export const editComment = async ({
   await updateDoc(commentsRef, comment).then(() =>
     updateDoc(commentsRef, newComment),
   );
+};
+
+// firestore에서 'Users' 데이터 볼러 옴
+export const getUsersList = async () => {
+  const array: any[] = [];
+
+  const q = query(collection(db, 'Users'));
+
+  const querySnapshot = await getDocs(q);
+  querySnapshot.forEach((doc) =>
+    array.push({
+      ...doc.data(),
+    }),
+  );
+
+  return array;
 };
