+++ src/pages/signup/index.tsx
@@ -74,7 +74,7 @@ const SignUp = () => {
     };
 
     await updateDoc(doc(db, 'Users', email), updateUser);
-    alert('회원가입이 완료되었습니다.')
+    alert('회원가입이 완료되었습니다.');
     router.push('/');
   };
